+++ civicgrants/__init__.py
@@ -1,7 +1,6 @@
 """CivicGrants package."""
 
-__version__ = "0.1.1"
-
+__version__ = "1.0.0"